+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Set `--azureml` automatically when running on Azure ML
+
 ## Version 0.0.2 (development)
 
 - Created package: marian-tensorboard